requirements.txt

streamlit
pymupdf
pandas
fpdf2
matplotlib
